requirements.txt
python-telegram-bot==20.8
flask==3.0.2
stripe==8.5.0
requests==2.31.0